Automatically crops image using standard libraries
This is a python class, where the channel to be cropped can be chosen. To avoid accidentally cropping empty sections, it is first checked that atleast 12% of the cropped section is non-empty before saving them (by using a combination of entropy filter, Otsu's thresholding and binary filters to check for % of non-empty sections automatically).